refactor: dir name
